+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - 🔒️(backend) enforce display name setting on rename API
 - 🩹(backend) handle failed and aborted egresses
+- 🩹(frontend) notify participants when a recording fails or is aborted
 
 ## [1.31.0] - 2026-09-08